Remove misleading reference to "publisherId" cause it is ignored
The only way to override the `Publisher` value in the generated `AppxManifest.xml` is to provide a custom signing certificate. Merely setting the "publisherId" does not do anything. I've tried it a whole day. Only when I set a custom signing certificate my applications manifest looked like it should. And then setting "publisherId" to anything did not affect how the resulting manifest file looked.
So I propose to remove the value from the docs (and it should also be removed from the Cordova Windows plugin).