Add VoltSigner to tools (#1175)
I posted this to the dev mailing list on April 30. As there were no replies, I hope it's ok to move to the next step:

We’ve got a new tool we’d like to share with the Cordova community. It facilitates the creation of signing certificates for Android and iOS.

https://voltsigner.com

- It’s free
- No login is required
- We’re in the process of making of making the source public on GitHub
- Runs completely in the browser
- Generates Android and iOS .p12 files on a Windows or MacOS computers.
- Certificates are made 100% locally - passwords and files are not shared to the cloud.