Avoid Automatic Code Block for HTML on Events Page
It seems that the spacing between the HTML elements on this page caused an automatic code block when converted for render. That code block caused the page content to be mixed up including rendering a heap of the Markdown as straight text in a jumbled, single-paragraph mess.

This change simply removes some of the whitespace lines between the "td" elements so that the HTML Table summarising the events and their support is treated as a single HTML table and not the start of one with HTML example code after it.